+++ tests/Tests/PHPUnit/Traits/InstancesAccessorsTraitTest.php
@@ -4,6 +4,7 @@ namespace Tests\AvtoDev\DevTools\Tests\PHPUnit\Traits;
 
 use Tests\AvtoDev\DevTools\AbstractTestCase;
 use PHPUnit\Framework\ExpectationFailedException;
+use AvtoDev\DevTools\Tests\PHPUnit\AbstractLaravelTestCase;
 use SebastianBergmann\RecursionContext\InvalidArgumentException;
 use AvtoDev\DevTools\Tests\PHPUnit\Traits\InstancesAccessorsTrait;
 
@@ -31,5 +32,34 @@ class InstancesAccessorsTraitTest extends AbstractTestCase
 
         $this->assertEquals('foo', $this->getProperty($instance, 'property'));
         $this->assertEquals('bar', $this->callMethod($instance, 'method'));
+    }
+
+    /**
+     * @return void
+     */
+    public function testGetClosureHash()
+    {
+        $test_cases = [
+            function () {
+            },
+            function () {
+                return new class extends AbstractLaravelTestCase {
+                };
+            },
+            function () {
+                $this->resetCount();
+            },
+        ];
+
+        $this->assertNotEmpty($test_cases);
+
+        foreach ($test_cases as $test_case) {
+            $hash = $this->getClosureHash($test_case);
+
+            $this->assertGreaterThanOrEqual(8, \mb_strlen($hash));
+
+            // Second call for a closure
+            $this->assertEquals($hash, $this->getClosureHash($test_case));
+        }
     }
 }